Used shade plugin to create a runnable JAR file; renamed Main to App and created a Main file
