+++ .gitignore
@@ -36,3 +36,6 @@ build/
 
 ### Mac OS ###
 .DS_Store
+
+### Maven shade plugin ###
+dependency-reduced-pom.xml
